Little doc string

+++ habitat_baselines/rl/ddppo/policy/resnet_policy.py
@@ -163,7 +163,9 @@ class ResNetEncoder(nn.Module):
             rgb_observations = observations["rgb"]
             # permute tensor to dimension [BATCH x CHANNEL x HEIGHT X WIDTH]
             rgb_observations = rgb_observations.permute(0, 3, 1, 2)
-            rgb_observations = rgb_observations / 255.0  # normalize RGB
+            rgb_observations = (
+                rgb_observations.float() / 255.0
+            )  # normalize RGB
             cnn_input.append(rgb_observations)
 
         if self._n_input_depth > 0:
